+++ action.sh
@@ -126,15 +126,8 @@ perform_backup() {
     
     # Backup enabled modules list
     echo "Backing up enabled modules list"
-    # Detect KSU by checking for KSU modules directory
-    if [ -d "/data/adb/ksu/modules" ]; then
-        MODULE_DIR="/data/adb/ksu/modules"
-    elif [ -d "/data/adb/modules" ]; then
-        MODULE_DIR="/data/adb/modules"
-    else
-        # Fallback - shouldn't happen but just in case
-        MODULE_DIR="/data/adb/modules"
-    fi
+    # All modern root solutions (Magisk, KSU, KSU Next, Sukisu Ultra, APatch) use /data/adb/modules
+    MODULE_DIR="/data/adb/modules"
     
     > "$BACKUP_DIR/modules/module_list.txt"